GdalXmlNode::fmt(): use explicit VSIFree()
While reviewing the code, I was wondering if it wasn't leaking, but it
seems not in practice. However using CString::from_raw() seems a bit
dangerous and relies on non specified behaviour, as the documentation of
that function mentions 'Retakes ownership of a CString that was
transferred to C via CString::into_raw', which isn't the case here.
The fact that it worked in practice comes from the fact the default
deallocator used behind the scenes by CString is free(), and
that VSIFree() is also an alias for that.